Refactored 'Remnant' stuff into own module
