+++ src/Day22.hs
@@ -5,6 +5,7 @@ module Day22 where
 import qualified Data.Text as T
 
 import Cuboid (Cuboid(..), parseCuboid, volume)
+import Remnant (Remnant, augment, reduce)
 
 solvePuzzle :: String -> IO ()
 solvePuzzle filePath = do
@@ -40,35 +41,8 @@ parseLine line =
                     , parseCuboid cuboidStr
                     )
 
-{- | A `Cuboid` list of non-overlapping cuboids
-   | generated by recursively recursively performing a `RebootStep`
-   | upon all the cuboids.
--}
-type Remnant = [ Cuboid ]
-
 generateRemnant :: [RebootStep] -> Remnant
 generateRemnant rbs = [] -- TODO
-
-{- | return the first Cuboid and the remnant
-   | of the second Cuboid.
-   |
-   | Properties
-   |
-   | 1. The combined volumes of the remnant == combined volumes of the 1st cuboid and the 2nd cuboid
-   |    minus their intersection
--}
-augment :: Cuboid -> Cuboid -> Remnant
-augment source target = [] -- TODO
-
-{- | return the remnant of subtracting the first
-   | Cuboid from the second.
-   |
-   | Properties:
-   |
-   | 1. The combined volumes of the remnant == 2nd Cuboid's volume - the volume of the 2 cuboids' intersection
--}
-reduce :: Cuboid -> Cuboid -> Remnant
-reduce source target = [] -- TODO
 
 {- Selects the function to implement the combining of the source and target cuboids
 -}
